Force initialization of canvas to start time specified in manifest when canvas index changes
This resolves the issue with playlist items being skipped (#375) and may also resolve #385.
Playlist items were being skipped because the handleTimeUpdate method in VideoJSProgress was firing a second 'ended' event
while the player was in the middle of initializing the new playlist item.  This was happening because the start time of the
previous playlist item was still the currentTime which was greater than the end time of the initializing playlist item.
Setting the `fromStart` flag when calling initCanvas forces the player to set current time in the state to what is specified
in the manifest.  This shouldn't affect structure navigation clicks across canvases because that should be handled by the
`switchCanvas` method which passes along a value for `fromStart` instead of assuming a default.