Check if player is paused before firing ended events
I was not able to track down the root cause, but not all player
instances would be assigned `player.ended() === true` when we triggered
the `ended` event in the time update handler. This was mainly seen with
playlist items, but it did not seem to occur with all playlists. The
`pause` event, however, does seem to be triggering consistently and
assigning the proper values. When checking if we are at or past the end
of the video, we can prevent continuous state updates by also checking
if the player is already paused, making the assumption that this means
the code within the block has already run once.