Sanitize manifest url if params included in middle
This commit is to fix an issue found with Ramp embedded in Avalon. If
the request url from avalon has params on the end, Avalon feeds the
entire url into Ramp as
`https://example.com/endpoint?params/manifest.json`. This causes media
to fail to load because it is not a valid manifest url. By removing the
params from the middle of the url, we can get Ramp to request the proper
manifest url.